60-langgraph-agent-handoffs/pyproject.toml
[project]
name = "60-langgraph-agent-handoffs"
version = "0.1.0"
description = "Langgraph agent handoffs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.12",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
]
